refactor : refactor models
