--- article/models/__init__.py
@@ -1,5 +1,0 @@
-from .articles.article import Article
-from .articles.category import Category
-from .articles.mail_batch import MailBatch
-from .user_info.user_category import UserCategory
-from .user_info.user_schedule import UserSchedule
