+++ article_app/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class ArticleConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'article'
+    name = 'article_app'
