+++ article_app/models/base_model.py
