+++ article_app/models/user_category/_models.py
@@ -1,9 +1,9 @@
 from django.db import models
-from article.models.base_model import AbstractBaseModel
+from article_app.models.base_model import AbstractBaseModel
 
 class UserCategory(AbstractBaseModel):
     user_email = models.EmailField()
-    category = models.ForeignKey('article.Category', on_delete=models.CASCADE, related_name='user_categories')
+    category = models.ForeignKey('article_app.Category', on_delete=models.CASCADE, related_name='user_categories')
     is_activated = models.BooleanField(default=True)
     sent_mail_count = models.BigIntegerField(default=0)
