+++ article_app/serializers/articles/category_serializers.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from article.models.articles.category import Category
+from article_app.models.category._models import Category
 
 class CategorySerializer(serializers.ModelSerializer):
     class Meta:
