+++ article_app/urls.py
@@ -1,8 +1,8 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from article.views.user_info import UserCategoryViewSet,UserScheduleViewSet
-from article.views import HealthCheckView
-from article.views.article import CategoryViewSet,ArticleViewSet
+from article_app.views.user_info import UserCategoryViewSet,UserScheduleViewSet
+from article_app.views import HealthCheckView
+from article_app.views.article import CategoryViewSet,ArticleViewSet
 
 router = DefaultRouter()
 router.register(r'category', CategoryViewSet, basename='category')
