+++ batch_app/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class BatchConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'batch'
+    name = 'batch_app'
