+++ django_server/settings.py
@@ -37,8 +37,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework',
-    'article',
-    'batch',
+    'article_app',
+    'batch_app',
     'drf_yasg',
 ]
